scripts/requirements.txt
pdfminer.six
pdf2image
pytesseract
